Move the external config file
Move shared config file into a config folder. Added __DIR__ to path in to
support the use of project in vendor folder as root.
The default config file as a separate file is removed. Users are free to
create a backup of the file or come back to the repo to re-download. There
iss no clear benefit to loading 2 files and merge the arrays for every
operation.
The config file is also a JSON file and should have the correct
extension.