Update paths in JS to be relative to project folder
The paths in the JS file is now referencing to
config.path.filemanagerRoot to get to the file. This allows the
front-end code to continue to work even it's not in the root folder.
Move fileConnector in config to "path" instead of options so it sits
with all the other paths.